+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Constants/VirtualTutorialLessons.php
@@ -10,12 +10,14 @@ namespace GatoGraphQL\GatoGraphQL\Constants;
  */
 class VirtualTutorialLessons
 {
-    public final const IMPORTING_A_POST_FROM_WORDPRESS_RSS_FEED = 'importing-a-post-from-wordpress-rss-feed';
-    public final const IMPORTING_POSTS_FROM_A_CSV = 'importing-posts-from-a-csv';
-    public final const TRANSLATING_CLASSIC_EDITOR_POST_TO_A_DIFFERENT_LANGUAGE = 'translating-classic-editor-post-to-a-different-language';
+    public final const ADD_COMMENTS_BLOCK_TO_POST = 'add-comments-block-to-post';
     public final const BULK_TRANSLATING_CLASSIC_EDITOR_POSTS_TO_A_DIFFERENT_LANGUAGE = 'bulk-translating-classic-editor-posts-to-a-different-language';
     public final const FETCH_POST_LINKS = 'fetch-post-links';
-    public final const SEND_EMAIL_TO_ADMIN_ABOUT_POST = 'send-email-to-admin-about-post';
-    public final const ADD_COMMENTS_BLOCK_TO_POST = 'add-comments-block-to-post';
     public final const GENERATE_A_POST_FEATURED_IMAGE_USING_AI_AND_OPTIMIZE_IT = 'generate-a-post-featured-image-using-ai-and-optimize-it';
+    public final const IMPORTING_A_POST_FROM_WORDPRESS_RSS_FEED = 'importing-a-post-from-wordpress-rss-feed';
+    public final const IMPORTING_POSTS_FROM_A_CSV = 'importing-posts-from-a-csv';
+    public final const SEND_EMAIL_TO_ADMIN_ABOUT_POST = 'send-email-to-admin-about-post';
+    public final const TRANSLATING_CLASSIC_EDITOR_POST_TO_A_DIFFERENT_LANGUAGE = 'translating-classic-editor-post-to-a-different-language';
+    public final const TRANSLATING_POSTS_FOR_POLYLANG_AND_GUTENBERG = 'translating-posts-for-polylang-and-gutenberg';
+    public final const TRANSLATING_POSTS_FOR_POLYLANG_AND_CLASSIC_EDITOR = 'translating-posts-for-polylang-and-classic-editor';
 }